+++ GloryEngine/Engine/GloryCore/MeshData.h
@@ -9,15 +9,15 @@ namespace Glory
 	{
 	public:
 		MeshData();
-		MeshData(size_t vertexCount, size_t vertexSize, const std::vector<float>& vertices, size_t indexCount, const std::vector<uint32_t>& indices, const std::vector<AttributeType>& attributes);
-		MeshData(size_t vertexCount, size_t vertexSize, const float* vertices, size_t indexCount, const uint32_t* indices, const std::vector<AttributeType>& attributes);
-		MeshData(size_t vertexCount, size_t vertexSize, const float* vertices, const std::vector<AttributeType>& attributes);
+		MeshData(uint32_t vertexCount, uint32_t vertexSize, const std::vector<float>& vertices, uint32_t indexCount, const std::vector<uint32_t>& indices, const std::vector<AttributeType>& attributes);
+		MeshData(uint32_t vertexCount, uint32_t vertexSize, const float* vertices, uint32_t indexCount, const uint32_t* indices, const std::vector<AttributeType>& attributes);
+		MeshData(uint32_t vertexCount, uint32_t vertexSize, const float* vertices, const std::vector<AttributeType>& attributes);
 		virtual ~MeshData();
 
-		const size_t VertexCount() const;
-		const size_t IndexCount() const;
-		const size_t VertexSize() const;
-		const size_t AttributeCount() const;
+		const uint32_t VertexCount() const;
+		const uint32_t IndexCount() const;
+		const uint32_t VertexSize() const;
+		const uint32_t AttributeCount() const;
 		const float* Vertices() const;
 		const uint32_t* Indices() const;
 		const AttributeType* AttributeTypes() const;
@@ -27,8 +27,8 @@ namespace Glory
 		std::vector<float> m_Vertices;
 		std::vector<uint32_t> m_Indices;
 		std::vector<AttributeType> m_Attributes;
-		const size_t m_VertexCount;
-		const size_t m_IndexCount;
-		const size_t m_VertexSize;
+		const uint32_t m_VertexCount;
+		const uint32_t m_IndexCount;
+		const uint32_t m_VertexSize;
 	};
 }
