+++ GloryEngine/Engine/Modules/GloryOpenGLGraphics/GLMesh.h
@@ -8,7 +8,7 @@ namespace Glory
 	class GLMesh : public Mesh
 	{
     public:
-        GLMesh(size_t vertexCount, size_t indexCount, InputRate inputRate, size_t binding, size_t stride, const std::vector<AttributeType>& attributeTypes);
+        GLMesh(uint32_t vertexCount, uint32_t indexCount, InputRate inputRate, uint32_t binding, uint32_t stride, const std::vector<AttributeType>& attributeTypes);
         virtual ~GLMesh();
         virtual void CreateBindingAndAttributeData() override;
         virtual void Bind() override;
